Fixed IntelliSense and syntax highlighting
