+++ ipython-profile/ipython_config.py
@@ -1,4 +1,5 @@
 c = get_config()
-c.KernelManager.kernel_spec = [r"%s", "{connection_file}"]
+c.KernelManager.kernel_spec = [ "mono", r"%kexe", "{connection_file}"]
 c.Session.key = b''
 c.Session.keyfile = ''
+c.NotebookApp.extra_static_paths = [ r"%kfolder" ]
