+++ src/IfSharp.Kernel/IfSharpResources.resx
@@ -118,11 +118,11 @@
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
   <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
-  <data name="custom_css" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\..\ipython-profile\static\custom\custom.css;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
+  <data name="fsharp_css" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\..\ipython-profile\static\custom\fsharp.css;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
   </data>
-  <data name="custom_js" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\..\ipython-profile\static\custom\custom.js;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
+  <data name="kernel_js" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\..\ipython-profile\kernel.js;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
   </data>
   <data name="fsharp_js" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\..\ipython-profile\static\custom\fsharp.js;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
@@ -132,6 +132,18 @@
   </data>
   <data name="ipython_config" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\..\ipython-profile\ipython_config.py;System.Byte[], mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </data>
+  <data name="kernel_json" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\..\kernel-spec\kernel.json;System.Byte[], mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </data>
+  <data name="logo32File" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\..\kernel-spec\logo-32x32.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="logo64File" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\..\kernel-spec\logo-64x64.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="qtconsole_config" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\..\ipython-profile\ipython_qtconsole_config.py;System.Byte[], mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </data>
   <data name="webintellisense" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\..\ipython-profile\static\custom\webintellisense.js;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;utf-8</value>
