+++ src/IfSharpConsole/Program.cs
@@ -1,4 +1,5 @@
 ﻿using IfSharp.Kernel;
+using System.Globalization;
 
 namespace IfSharpConsole
 {
@@ -6,6 +7,8 @@ namespace IfSharpConsole
     {
         static void Main(string[] args)
         {
+            CultureInfo.DefaultThreadCurrentCulture = CultureInfo.InvariantCulture;
+            CultureInfo.DefaultThreadCurrentUICulture = CultureInfo.InvariantCulture;
             App.Start(args);
         }
     }
